--- frontend/src/homes/onboards/components/RollupSummaryPanel.jsx
@@ -1,107 +1,0 @@
-import { useState, useEffect } from "react";
-import { getStatus, calcBatch } from "@/apis/rollup";
-import { showDefaultAlert } from "@components/UI/ServiceAlert";
-
-const RollupSummaryPanel = ({ batchId, onCalculated }) => {
-  const [statusInfo, setStatusInfo] = useState(null);
-  const [calculating, setCalculating] = useState(false);
-
-  useEffect(() => {
-    if (!batchId) return;
-    fetchStatus();
-    // In a real scenario, this might poll the status.
-    // const timer = setInterval(fetchStatus, 10000);
-    // return () => clearInterval(timer);
-  }, [batchId]);
-
-  const fetchStatus = async () => {
-    try {
-      const res = await getStatus(batchId);
-      if (res?.data) {
-        setStatusInfo(res.data);
-      }
-    } catch (e) {
-      console.error(e);
-      // Fallback mock
-      setStatusInfo({
-        batchId,
-        status: 'PENDING',
-        totalSubsidiaries: 2,
-        completedSubsidiaries: 1,
-        progressPercent: 50
-      });
-    }
-  };
-
-  const handleCalc = async () => {
-    setCalculating(true);
-    try {
-      const res = await calcBatch(batchId);
-      if (res?.status) {
-        showDefaultAlert("성공", "롤업 계산이 완료되었습니다.", "success");
-        fetchStatus();
-        onCalculated?.();
-      } else {
-        showDefaultAlert("오류", "계산에 실패했습니다.", "error");
-      }
-    } catch (e) {
-      console.error(e);
-      // mock success
-      setTimeout(() => {
-        setStatusInfo(prev => ({ ...prev, status: 'CALCULATED', progressPercent: 100, completedSubsidiaries: prev.totalSubsidiaries }));
-        showDefaultAlert("성공", "롤업 계산이 완료되었습니다.", "success");
-        onCalculated?.();
-        setCalculating(false);
-      }, 1000);
-    }
-  };
-
-  if (!statusInfo) return null;
-
-  return (
-    <div style={{
-      background: '#fff',
-      border: '1px solid #e2e8f0',
-      borderRadius: '8px',
-      padding: '16px',
-      marginBottom: '24px',
-      boxShadow: '0 1px 2px rgba(0,0,0,0.02)',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'space-between'
-    }}>
-      <div style={{ display: 'flex', flexDirection: 'column', gap: '4px' }}>
-        <h3 style={{ margin: 0, fontSize: '0.95rem', color: '#1e293b' }}>
-          자회사 G0 데이터 롤업 현황
-        </h3>
-        <span style={{ fontSize: '0.85rem', color: '#64748b' }}>
-          수집 완료: {statusInfo.completedSubsidiaries} / {statusInfo.totalSubsidiaries} 개사 
-          ({statusInfo.progressPercent}%)
-        </span>
-      </div>
-      
-      <div style={{ display: 'flex', alignItems: 'center', gap: '16px' }}>
-        <div style={{ width: '150px', height: '8px', background: '#e2e8f0', borderRadius: '4px', overflow: 'hidden' }}>
-          <div style={{ width: `${statusInfo.progressPercent}%`, height: '100%', background: '#3b82f6', transition: 'width 0.3s' }}></div>
-        </div>
-        <button 
-          style={{ 
-            padding: '8px 16px', 
-            background: statusInfo.status === 'CALCULATED' ? '#10b981' : '#1d4ed8', 
-            color: '#fff', 
-            border: 'none', 
-            borderRadius: '4px',
-            cursor: calculating ? 'not-allowed' : 'pointer',
-            opacity: calculating ? 0.7 : 1
-          }}
-          onClick={handleCalc}
-          disabled={calculating}
-        >
-          {calculating ? "계산 중..." : statusInfo.status === 'CALCULATED' ? "계산 완료" : "롤업 계산 실행"}
-        </button>
-      </div>
-    </div>
-  );
-};
-
-export default RollupSummaryPanel;
